# Call VZEROALL for a 21% benchmark improvement