```
goos: linux
goarch: amd64
pkg: github.com/bwesterb/go-and
cpu: 13th Gen Intel(R) Core(TM) i9-13900
          │     asm     │          with_vzeroall/asm          │
          │   sec/op    │   sec/op     vs base                │
And-32      631.2n ± 2%   518.4n ± 1%  -17.87% (p=0.000 n=10)
Or-32       626.4n ± 1%   515.0n ± 1%  -17.79% (p=0.000 n=10)
Xor-32      631.6n ± 1%   518.6n ± 1%  -17.88% (p=0.000 n=10)
AndNot-32   635.2n ± 2%   517.8n ± 1%  -18.49% (p=0.000 n=10)

          │     asm      │          with_vzeroall/asm           │
          │     B/s      │     B/s       vs base                │
And-32      47.21Gi ± 2%   57.49Gi ± 1%  +21.77% (p=0.000 n=10)
Or-32       47.58Gi ± 1%   57.87Gi ± 1%  +21.63% (p=0.000 n=10)
Xor-32      47.19Gi ± 1%   57.47Gi ± 1%  +21.78% (p=0.000 n=10)
AndNot-32   46.92Gi ± 2%   57.56Gi ± 1%  +22.68% (p=0.000 n=10)
```